refactor: 投稿の配送Jobの改善
